Address perf issue in https://github.com/microsoft/onnxruntime/issues/7212 

Problem: 
- AlgoSearchWorkspaceSize was set to 32MB. The workspace size is too small that cudnnFindConvolutionForwardAlgorithmEx couldn't find the optimal algorithm for Conv forward. (The final algo end up requiring 169 MB workspace)
- In Conv1D case, if the input shape is [N, C, D], ORT is padding it to [N, C, D, 1]. This is not as efficient as padding to [N, C, 1, D]. 

#### cudnn log after the fix
```
I! CuDNN (v8003) function cudnnConvolutionForward() called:
i!     handle: type=cudnnHandle_t; streamId=0xf2b1040;
i!     alpha: type=CUDNN_DATA_FLOAT; val=1.000000;
i!     xDesc: type=cudnnTensorDescriptor_t:
i!         dataType: type=cudnnDataType_t; val=CUDNN_DATA_HALF (2);
i!         nbDims: type=int; val=4;
i!         dimA: type=int; val=[32,896,1,832];
i!         strideA: type=int; val=[745472,832,832,1];
i!     xData: location=dev; addr=0x7f9498000000;
i!     wDesc: type=cudnnFilterDescriptor_t:
i!         dataType: type=cudnnDataType_t; val=CUDNN_DATA_HALF (2);
i!         vect: type=int; val=0;
i!         nbDims: type=int; val=4;
i!         dimA: type=int; val=[1792,896,1,19];
i!         format: type=cudnnTensorFormat_t; val=CUDNN_TENSOR_NCHW (0);
i!     wData: location=dev; addr=0x7f946c000000;
i!     convDesc: type=cudnnConvolutionDescriptor_t:
i!         mode: type=cudnnConvolutionMode_t; val=CUDNN_CROSS_CORRELATION (1);
i!         dataType: type=cudnnDataType_t; val=CUDNN_DATA_FLOAT (0);
i!         mathType: type=cudnnMathType_t; val=CUDNN_TENSOR_OP_MATH (1);
i!         reorderType: type=int; val=0;
i!         arrayLength: type=int; val=2;
i!         padA: type=int; val=[0,9];
i!         strideA: type=int; val=[1,2];
i!         dilationA: type=int; val=[1,1];
i!         groupCount: type=int; val=1;
i!     algo: type=cudnnConvolutionFwdAlgo_t; val=CUDNN_CONVOLUTION_FWD_ALGO_IMPLICIT_PRECOMP_GEMM (1);
i!     workSpace: location=dev; addr=0x7f9408000000;
i!     workSpaceSizeInBytes: type=size_t; val=169747984;
i!     beta: type=CUDNN_DATA_FLOAT; val=0.000000;
i!     yDesc: type=cudnnTensorDescriptor_t:
i!         dataType: type=cudnnDataType_t; val=CUDNN_DATA_HALF (2);
i!         nbDims: type=int; val=4;
i!         dimA: type=int; val=[32,1792,1,416];
i!         strideA: type=int; val=[745472,416,416,1];
i!     yData: location=dev; addr=0x7f9494000000;
i! Time: 2021-04-15T08:21:17.252263 (0d+0h+0m+2s since start)
i! Process=40117; Thread=40117; GPU=0; Handle=0x3e070ba0; StreamId=0xf2b1040.
```


Profiling result: 

#### Pad to [N, C, D, 1], total time is 1.13s for 100 steps

```
Profiling result:
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   42.53%  1.28268s       400  3.2067ms  1.6640us  13.918ms  [CUDA memcpy HtoD]
                   36.06%  1.08757s       100  10.876ms  10.868ms  10.895ms  void xmma_new::gemm::kernel<xmma_new::implicit_gemm::fprop_indexed::Kernel_traits<xmma_new::Volta_hmma_fp32_traits, xmma_new::Cta_tile<xmma_new::Volta, int=128, int=128, int=64, int=2, int=2, int=2, int=1, int=1>, xmma_new::implicit_gemm::fprop_indexed::Gmem_tile_a_t<xmma_new::Volta_hmma_fp32_traits, xmma_new::Cta_tile<xmma_new::Volta, int=128, int=128, int=64, int=2, int=2, int=2, int=1, int=1>, xmma_new::implicit_gemm::Input_related<int=0, int=0, int=0, bool=0>, int=16, bool=0, xmma_new::implicit_gemm::fprop_indexed::Gmem_tile_base_a<xmma_new::Volta_hmma_fp32_traits, xmma_new::Cta_tile<xmma_new::Volta, int=128, int=128, int=64, int=2, int=2, int=2, int=1, int=1>, xmma_new::implicit_gemm::Input_related<int=0, int=0, int=0, bool=0>, int=16, xmma_new::Row, int=64, int=128>>, xmma_new::implicit_gemm::fprop_indexed::Gmem_tile_c_t<xmma_new::Volta_hmma_fp32_traits, xmma_new::Cta_tile<xmma_new::Volta, int=128, int=128, int=64, int=2, int=2, int=2, int=1, int=1>, int=16, xmma_new::Fragment_c<xmma_new::Volta_hmma_fp32_traits, xmma_new::Cta_tile<xmma_new::Volta, int=128, int=128, int=64, int=2, int=2, int=2, int=1, int=1>, bool=1>>, xmma_new::implicit_gemm::Input_related<int=0, int=0, int=0, bool=0>, int=1>>(xmma_new::Volta_hmma_fp32_traitsParams)
                   19.93%  601.03ms       100  6.0103ms  5.8734ms  8.7697ms  [CUDA memcpy DtoH]
                    1.06%  32.003ms       200  160.01us  125.79us  195.42us  void nchwToNhwcKernel<__half, __half, float, bool=1, bool=0, cudnnKernelDataType_t=0>(int, int, int, int, __half const *, __half*, float, float)
                    0.39%  11.743ms       100  117.43us  116.77us  119.04us  void nhwcToNchwKernel<__half, __half, float, bool=1, bool=0, cudnnKernelDataType_t=0>(int, int, int, int, __half const *, __half*, float, float)
                    0.02%  704.95us       400  1.7620us  1.6630us  3.4880us  [CUDA memset]
```

#### Pad to [N, C, 1, D], total time is 1.0605 sec for 100 steps
```
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   43.11%  1.25958s       400  3.1490ms  1.6320us  8.2548ms  [CUDA memcpy HtoD]
                   35.20%  1.02850s       100  10.285ms  10.272ms  10.297ms  volta_fp16_s884cudnn_fp16_256x128_ldg8_relu_f2f_exp_large_nhwc2nchw_tn_v1
                   20.55%  600.52ms       100  6.0052ms  5.8856ms  8.0205ms  [CUDA memcpy DtoH]
                    1.10%  32.008ms       200  160.04us  125.82us  195.01us  void nchwToNhwcKernel<__half, __half, float, bool=1, bool=0, cudnnKernelDataType_t=0>(int, int, int, int, __half const *, __half*, float, float)
                    0.02%  704.73us       400  1.7610us  1.6640us  3.4880us  [CUDA memset]
                    0.01%  337.25us       100  3.3720us  2.5280us  4.8640us  cask_cudnn::computeOffsetsKernel(cask_cudnn::ComputeOffsetsParams)
```