@vadimkantorov 2021-04-19
Is it expected that convDesc.dataType is CUDNN_DATA_FLOAT after the fix, and not CUDNN_DATA_HALF?