@vadimkantorov 2021-04-19
I asked the same question in PyTorch: https://github.com/pytorch/pytorch/issues/55183#issuecomment-813321237

The question is what's semantics of convDesc.dataType? would it be correct to set it to CUDNN_DATA_HALF? would it lead to faster results?