@vadimkantorov 2021-04-22
Hope there would be a 1.7.3 release with this inside :) For now we'll test the nightly version